Support delaying mock responses by a fixed number of milliseconds (#103)
This feature might be useful for predicting the response time
of a handler. When using the Debug(...) helper the total time of the handler
is printed to the console. It might be desirable to only use this during
experiments and to not run this config in CI. This is why the global
option must be enabled to turn this on.